requirements.txt
anthropic
flask
gunicorn
openai
pandas
PyMuPDF
pypdf
python-dotenv
requests
scikit-learn
twilio
